nautilus/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spy-orb-nautilus"
version = "0.1.0"
description = "SPY 0DTE ORB strategy on NautilusTrader + moomoo OpenD"
requires-python = ">=3.12"

dependencies = [
    "nautilus_trader>=1.225.0",
    "nautilus-futu>=0.4.2",
    "yfinance>=0.2",
    "pandas>=2.0",
    "python-dotenv",
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-asyncio>=0.23"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
